+++ src/Shopgate/Extended/Flysystem/ExtendedPluginApi.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Shopgate\Shopware\Shopgate\Extended\Core;
+namespace Shopgate\Shopware\Shopgate\Extended\Flysystem;
 
 use League\Flysystem\FileNotFoundException;
 use League\Flysystem\FilesystemInterface;